+++ src/main/java/org/heigit/ohsome/filter/GeometryFilterArea.java
@@ -1,0 +1,13 @@
+package org.heigit.ohsome.filter;
+
+import javax.annotation.Nonnull;
+import org.heigit.bigspatialdata.oshdb.util.geometry.Geo;
+
+/**
+ * A filter which checks the area of OSM feature geometries.
+ */
+public class GeometryFilterArea extends GeometryFilter {
+  GeometryFilterArea(@Nonnull ValueRange range) {
+    super(range, GeometryMetricEvaluator.fromLambda(Geo::areaOf, "area"));
+  }
+}